python/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
# JWT via stdlib hmac+hashlib (no cryptography dep needed)
bcrypt>=4.0.0
pydantic>=2.5.0
streamlit>=1.30.0
requests>=2.31.0
python-dotenv>=1.0.0
